Conflicts:
	cpu/samd21/samd21_linkerscript.ld
	drivers/at86rf231/at86rf231_tx.c

+++ sys/oneway-malloc/oneway-malloc.c
@@ -60,7 +60,7 @@ void __attribute__((weak)) *realloc(void *ptr, size_t size)
     }
 }
 
-void __attribute__((weak)) *calloc(int size, size_t cnt)
+void __attribute__((weak)) *calloc(size_t size, size_t cnt)
 {
     void *mem = malloc(size * cnt);
     if (mem) {
